Rename b0 to gre
